+++ src/lib/paths.test.ts
@@ -1,7 +1,9 @@
+import { APP_NAME } from '@src/lib/constants'
 import { moduleFsViaModuleImport, StorageName } from '@src/lib/fs-zds'
 import { beforeAll, expect, describe, it } from 'vitest'
 import {
   getFilePathRelativeToProject,
+  getProjectRelativeFilePath,
   getRouterSearchFromRequestUrl,
   parseProjectRoute,
   toProjectRelativePath,
@@ -108,6 +110,20 @@ describe('testing web-safe project paths', () => {
         '/some/path/Simple Box/parts/generated/nested-part.kcl'
       )
     ).toEqual('parts/generated/nested-part.kcl')
+  })
+
+  it('should return the app name when there is no file', () => {
+    expect(getProjectRelativeFilePath()).toEqual(APP_NAME)
+  })
+
+  it('should return the file name when a relative path cannot be derived', () => {
+    expect(
+      getProjectRelativeFilePath(undefined, {
+        name: 'nested-part.kcl',
+        path: '',
+        children: null,
+      })
+    ).toEqual('nested-part.kcl')
   })
 })
