+++ src/lib/paths.ts
@@ -1,7 +1,8 @@
 import type { Configuration } from '@rust/kcl-lib/bindings/Configuration'
-import { ARCHIVE_DIR, IS_PLAYWRIGHT_KEY } from '@src/lib/constants'
+import { APP_NAME, ARCHIVE_DIR, IS_PLAYWRIGHT_KEY } from '@src/lib/constants'
 import fsZds from '@src/lib/fs-zds'
 
+import type { FileEntry, Project } from '@src/lib/project'
 import { err } from '@src/lib/trap'
 import type { DeepPartial } from '@src/lib/types'
 import type { ModuleType } from '@src/lib/wasm_lib_wrapper'
@@ -258,6 +259,24 @@ export function toProjectRelativePath(
   filePath: string
 ): string {
   return toWebSafePath(fsZds.relative(projectPath, filePath))
+}
+
+export function getProjectRelativeFilePath(
+  project?: Project,
+  file?: FileEntry
+): string {
+  if (!file) {
+    return APP_NAME
+  }
+
+  if (project?.path && file.path) {
+    const relativeFilePath = toProjectRelativePath(project.path, file.path)
+    if (relativeFilePath) {
+      return relativeFilePath
+    }
+  }
+
+  return toWebSafePath(file.name || APP_NAME)
 }
 
 /**
